Changed structure of log functions. Added new alog
Completely changed the structure of the log functions of the Riccati Bessel functions to separate them into 3 functions for A3, A4, and B. This also removed the bug where a for loop was being computed every time the function was called. The values are now pre computed before the main loop. 

Added a new file: debyeSeries_shen.py. This code follows the algorithm in Shen & Wang (2010, Applied Optics , 49, 2422). Referenced much of the code in Ryo Tazaki (2020/June/19). This version of the new function matches MiePlot from Phillip Laven exactly. Moving forward this code will be used as a base to implement the multilayered sphere. This logarithm is much faster and more stable as it does not require and direct evaluation of any bessel functions. They are all computed using recursion.