fixed bug with T^j_n calculation
There was a bug in the code with how T^j_n was being calculated. Shen and Wang (2010 Appl Opt) equation 11a and 11b show Tn being computed with a leading "m" value (index of refraction). The paper is for a single layer sphere and m describes the relative index of refraction. The code mistakenly used the absolution index of refraction causing T_n, and Q_n to blow up to inf. This is now fixed in all 3 scripts.